Collect all frontmatter before processing files for output.
This change has a cost to startup but it will permit extensions to
consider relationships between files before rendering.

The blog extension is the driving use case (for #79). handroll should
be able to discover all the blog posts to consider their order in
relation to each other. This will permit functionality like a previous
and next link for post pages.